fix: move CallApi to Core project
