+++ doc/WalkingTec.Mvvm.Doc/SetupFiles/Mvc/HomeController.txt
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.DependencyInjection;
 using WalkingTec.Mvvm.Core;
+using WalkingTec.Mvvm.Core.Extensions;
 using WalkingTec.Mvvm.Core.Auth;
 using WalkingTec.Mvvm.Mvc;
 using $vmns$.HomeVMs;
